% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanVector.R
\name{CleanVector}
\alias{CleanVector}
\title{Clean Vector}
\usage{
CleanVector(Vector, Type)
}
\arguments{
\item{Vector}{Vector with the scrapped values}

\item{Type}{Type of information stored in the vector}
}
\value{
A vector with all the variables in the correct format
}
\description{
Clean Vector
}
\examples{
/dontrun{
Area <-CleanVector(PRE_CARACT, "Area")
}
}
