% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDatabase.R
\name{CreateDatabase}
\alias{CreateDatabase}
\title{Create Database}
\usage{
CreateDatabase(Cartier, Property, Price, Price_Group, Area, Area_Group, Pieces,
  Chambres)
}
\arguments{
\item{Cartier}{Cartier Array}

\item{Property}{Property Array}

\item{Price}{Price Array}

\item{Price_Group}{Price group array}

\item{Area}{Area Array}

\item{Area_Group}{Area group array}

\item{Pieces}{Pieces Array}

\item{Chambres}{Chambres Array}
}
\value{
Data frame with the input information
}
\description{
Create Database
}
\examples{
/dontrun{
CreateData(Cartier, Property, Price, Area, Area_Group, Pieces, Chambres)
}
}
