% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixParameters.R
\name{FixParameters}
\alias{FixParameters}
\title{Fix Parameters}
\usage{
FixParameters(Value, Type)
}
\arguments{
\item{Value}{Array with the information scraped that needs to be fix}

\item{Type}{Type of information in the array ("Price", "Piece", "Chambre", "Parameters")}
}
\value{
a clean value of the information scraped
}
\description{
Fix Parameters
}
\examples{

/dontrun{
 FixParameters(AreaScraped, "Area")
}


}
