% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPages.R
\name{GetPages}
\alias{GetPages}
\title{Get pages}
\usage{
GetPages(WEB_PAGE)
}
\arguments{
\item{WEB_PAGE}{Selonger link}
}
\value{
the number of pages the link contains
}
\description{
Get pages
}
\examples{
\dontrun{
  Link<-http://www.seloger.com/list.htm?idtt=2&naturebien=1,2,4&idtypebien=1,2&ci=750056&tri=initial&surfacemin=5&surfacemax=9
  WEB_PAGE<-read_html(Link)
  GetPages(WEB_PAGES)
}
}
