% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupsArea.R
\name{GroupsArea}
\alias{GroupsArea}
\title{Groups Areas}
\usage{
GroupsArea(Area)
}
\arguments{
\item{Area}{the value of the area}
}
\value{
the group to which the area belongs to
}
\description{
Groups Areas
}
\examples{
\dontrun{
 GroupsArea(Area_Scraped)
}
}
