% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupsPrice.R
\name{GroupsPrice}
\alias{GroupsPrice}
\title{Groups Prices}
\usage{
GroupsPrice(Price)
}
\arguments{
\item{Price}{Price Vector}
}
\value{
the group to which the price belongs to
}
\description{
Groups Prices
}
\examples{
/dontrun{
 GroupsArea(Area_Scraped)
}
}
