% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapsInfo.R
\name{ScrapsInfo}
\alias{ScrapsInfo}
\title{ScrapsInfo}
\usage{
ScrapsInfo(link)
}
\arguments{
\item{link}{The link of the selonger pages that is suposed to be scrap}
}
\value{
a dataframe with the information scraped from the web page
}
\description{
ScrapsInfo
}
\examples{
\dontrun{
  Link<-http://www.seloger.com/list.htm?idtt=2&naturebien=1,2,4&idtypebien=1,2&ci=750056&tri=initial&surfacemin=5&surfacemax=9
  df<-ScrapsInfo(Link)
}
}
